requirements.txt
alembic==1.14.1
aniso8601==10.0.1
asttokens==3.0.0
backcall==0.2.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.8.2
certifi==2025.4.26
click==8.1.8
decorator==5.2.1
distlib==0.3.9
exceptiongroup==1.3.0
executing==2.2.0
Faker==35.2.2
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
geographiclib==1.52
geopy==2.2.0
greenlet==3.1.1
gunicorn==20.1.0
h11==0.16.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
pipenv==2024.4.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.3.0
pytest==8.3.5
python-dateutil==2.8.2
python-dotenv==0.19.0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
SQLAlchemy-serializer==1.4.12
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.13.2
virtualenv==20.31.2
virtualenv-clone==0.5.7
wcwidth==0.2.13
Werkzeug==3.0.6
wsproto==1.2.0
zipp==3.20.2
